V1.2.3 - Config Options
Implemented various configuration options for Ancient Tools, including the ability to disable bark bread and salves. Other options can be seen in the config file that generates on first load of this version.

Carry Capacity 'Carryable' behaviour is now only attached to bark baskets when Carry Capacity is actually installed.