V1.3.3
Fixed a crash that occurred with the adze when it was used on blocks that do not have attributes.

Added a hack-y compatibility patch that should make it possible for raw hides to be filled by right-clicking water, as intended, when Primitive Survival is also installed.

Vanilla Variants buckets can now be used in the crafting recipe to create raw hide sacks.

Made some modifications to the code that controls raw hide sack creation. They should now be made consistently when raw hide is interacted on water. Also, gave it a sound!

Fixed an issue where bark basket recipes couldn't be navigated to by clicking the bark basket on the bark handbook page.

Big thanks to Craluminum for pointing me in the right direction for several of these fixes.