V1.3.6
The mod can now be enjoyed in Japanese, courtesy of a translation offered by Macoto Hino!

Beeswax and fat patches should no longer overwrite the vanilla game attributes. This should allow the items to be placed on shelves again.
This should also fix compatibility issues with From Golden Combs on my end. However, a patch needs to be made on their end too to prevent overwriting of the Ancient Tools salve creation attribute.

in-container strings in language files have been duplicated to compensate for changes made in the game engine.