Ancient Tools_V1.4.6
Compiled the mod with VS RC-5 to prevent crashes that were happening with a 'Method Not Found' error with TesselateShape() that occurred when AT 1.4.5 was used on RC-5

Fixed an issue where sheep inventories were not properly increased in size.

Fixed an issue that prevented raw hides from being filled by clicking a water source.

Removed the UnstableFalling behavior from bark baskets to allow them to be placed anywhere.

Removed duplicate entries in all lang files.