Ancient Tools V1.5.23 + V1.5.24
Fixed an issue with the split logs that caused game sounds to disappear. Thanks to Agzil and Oscar for the detailed context of the issue, it helped immensely.

Provided an attribute on the SplitLog that allows other mods that add wood variants to supply drops for the wedge smacking process. This does nothing different on its own. Floral Zones mods will be getting updates to patch this in and solve existing crashes.

From 1.5.25/1.6.0: Fixed the Harmony for Fireproof Fuel patch to return true if no fireproof fuel was found. Vanilla torch now also uses this and was returning false, causing them to stay lit in rainy weather conditions.