Ancient Tools V1.6.0-pre.2 & pre.3
Pre-2:
Fixed a crash that occurred when a player would smash the bark stack with a tool and the bark stack was not a multiple of 4

Fixed a crash that occurred when interacting with logs that have no directional resin block

Fixed a brigadine armor recipe error

Added a sound effect for bark stack smashing

Pre-3:
Fixed an issue that prevented some in-world block/item conversions under certain conditions.

Fixed an issue that prevented the log barrel creation from being processed on the server.

Fixed an issue that caused scraping for resin to produce a debarked log instead.

Added unique sounds to the primitive barrel.