Ancient Tools V1.6.0-pre.5 & pre.6
pre.5
Fixed an issue where the final stage of crafting pitch would result in a '?' block or a pitch pot that cannot be interacted with.

Fixed an issue where halite stone and marble stones could not be ground down in the mortar.

Fixed an issue where right clicking logs with an empty hand would throw an exception.

pre.6
Fixed a crash that would occur in BlockBehaviorCarveLogBarrel when interacting with logs when other behaviors were being processed. Properly, this time, hopefully.

Pitch can once again be used to seal crocks.

Fixed an issue where chopping bark stacks with an axe would result in animation getting stuck and no further interaction was possible.

Removed ball pitch as a possible sealer for log barrels in the block interaction UI. It was only ever intended to accept the stick variant.